+++ inst/htmlwidgets/RacViewer/lib/functions/optimization.js
@@ -36,7 +36,8 @@ Racmacs.utils.titerType = function(titer) {
 Racmacs.utils.inc_base = function(
     map_dist,
     table_dist,
-    titer_type
+    titer_type,
+    dilution_stepsize
     ) {
 
     var ibase;
@@ -54,7 +55,7 @@ Racmacs.utils.inc_base = function(
       break;
     case 2:
       // Less than titer
-      x = table_dist - map_dist + 1;
+      x = table_dist - map_dist + dilution_stepsize;
       ibase = (10*x*x*Racmacs.utils.d_sigmoid(x) + 2*x*Racmacs.utils.sigmoid(x)) / map_dist;
       break;
     case 3:
@@ -74,7 +75,8 @@ Racmacs.utils.inc_base = function(
 Racmacs.utils.ptStress = function(
         map_dist,
         table_dist,
-        titer_type
+        titer_type,
+        dilution_stepsize
     ) {
 
     var x;
@@ -87,7 +89,7 @@ Racmacs.utils.ptStress = function(
       break;
     case 2:
       // Less than titer
-      x = table_dist - map_dist + 1;
+      x = table_dist - map_dist + dilution_stepsize;
       stress = Math.pow(x,2)*Racmacs.utils.sigmoid(x);
       break;
     case 3:
@@ -107,7 +109,8 @@ Racmacs.utils.ptStress = function(
 Racmacs.utils.ptResidual = function(
         map_dist,
         table_dist,
-        titer_type
+        titer_type,
+        dilution_stepsize
     ) {
 
     var x;
@@ -120,7 +123,7 @@ Racmacs.utils.ptResidual = function(
       break;
     case 2:
       // Less than titer
-      x = table_dist - map_dist + 1;
+      x = table_dist - map_dist + dilution_stepsize;
       residual = x*Racmacs.utils.sigmoid(x);
       break;
     case 3:
@@ -150,6 +153,7 @@ Racmacs.Optimizer = class RacOptimizer {
        this.num_sr           = args.sr_coords.length;
        this.titertype_matrix = args.titertypes;
        this.tabledist_matrix = args.tabledists;
+       this.dilutionstepsize = args.dilutionstepsize;
        this.mapdist_matrix = new Array(this.num_ags);
        for (var i=0; i<this.num_ags; i++) {
           this.mapdist_matrix[i] = new Array(this.num_sr);
@@ -232,7 +236,8 @@ Racmacs.Optimizer = class RacOptimizer {
           var ibase = Racmacs.utils.inc_base(
             this.mapdist_matrix[ag][sr],
             this.tabledist_matrix[ag][sr],
-            this.titertype_matrix[ag][sr]
+            this.titertype_matrix[ag][sr],
+            this.dilutionstepsize
           );
 
           // Now calculate the gradient for each coordinate
@@ -270,7 +275,8 @@ Racmacs.Optimizer = class RacOptimizer {
                 stress += Racmacs.utils.ptStress(
                     this.mapdist_matrix[ag][sr],
                     this.tabledist_matrix[ag][sr],
-                    this.titertype_matrix[ag][sr]
+                    this.titertype_matrix[ag][sr],
+                    this.dilutionstepsize
                 );
     
             }
@@ -321,7 +327,8 @@ Racmacs.Viewer.prototype.relaxMap = function(maxsteps) {
         tabledists : tabledists,
         titertypes : titertypes,
         fixed_ags : fixed_ags,
-        fixed_sr : fixed_sr
+        fixed_sr : fixed_sr,
+        dilutionstepsize : this.data.dilutionstepsize_cache
     });
 
     this.optimizer.update_map_dist_matrix();